% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri.R
\name{ComputeIriIndex}
\alias{ComputeIriIndex}
\title{Computes the IRI index from time series}
\usage{
ComputeIriIndex(ts)
}
\arguments{
\item{ts}{Input time series}
}
\description{
Computes the IRI index from time series
}
\examples{
ts1 <- rnorm(10)
iri <- ComputeIriIndex(ts)
}
\references{
some paper
}

