% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rroad-package.R
\docType{data}
\name{iri_coef_250mm}
\alias{iri_coef_250mm}
\title{Coeficients for computing IRI for 250mm segments}
\format{An object of class \code{list} of length 3.}
\usage{
iri_coef_250mm
}
\description{
Coeficients for computing IRI for 250mm segments
}
\keyword{datasets}

