% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iri.R
\name{CalculateIRI}
\alias{CalculateIRI}
\title{Computes the IRI for fixed length segments (e.g. 100 m segments) given a road profile}
\usage{
CalculateIRI(profile, sample.interval, segment.length = 100)
}
\arguments{
\item{profile}{Road profile (as numeric vector) whose IRI is to be calculated.}

\item{sample.interval}{Distance (in mm) between two samples of the given profile.}

\item{segment.length}{Distance (in m) for which the IRI is to be calculated. Default is 100 m.}
}
\value{
Calculated IRI (as numeric vector) of the given profile.
}
\description{
Computes the IRI for fixed length segments (e.g. 100 m segments) given a road profile
}
\examples{
iri <- CalculateIRI(profile, 10)
}

